feat(3.8.1): feat: h2DB支持 feat: 数据初始化初步支持(H2) feat: mail消息开关 refactor: shici, yiyan TODO 进一步优化数据初始化， TODO 文档 TODO 进一步开关控制 TODO 代码简洁度
