+++ src/main/java/com/bingchunmoli/api/interceptor/IpInterceptor.java
@@ -1,9 +1,9 @@
 package com.bingchunmoli.api.interceptor;
 
 import cn.hutool.extra.servlet.ServletUtil;
+import com.bingchunmoli.api.utils.RedisUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
@@ -20,14 +20,17 @@ import java.util.concurrent.atomic.AtomicInteger;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-@ConditionalOnBean(RedisTemplate.class)
 public class IpInterceptor implements HandlerInterceptor {
     private final RedisTemplate<String, Object> redisTemplate;
+    private final RedisUtil redisUtil;
 
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
 //        ServletContext servletContext = request.getServletContext();
+        if (redisUtil.isNotEnable()) {
+            return true;
+        }
         String requestURI = request.getRequestURI();
         String clientIP = ServletUtil.getClientIP(request);
         StringBuffer stringBuffer = new StringBuffer();
