+++ src/main/java/com/bingchunmoli/api/shici/service/impl/ShiCiServiceImpl.java
@@ -5,6 +5,7 @@ import com.bingchunmoli.api.bean.ApiConstant;
 import com.bingchunmoli.api.shici.bean.ShiCi;
 import com.bingchunmoli.api.shici.mapper.ShiCiMapper;
 import com.bingchunmoli.api.shici.service.IShiCiService;
+import com.bingchunmoli.api.utils.RedisUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.core.RedisTemplate;
@@ -26,9 +27,13 @@ import java.util.Random;
 public class ShiCiServiceImpl extends ServiceImpl<ShiCiMapper, ShiCi> implements IShiCiService {
     private final ShiCiMapper shiCiMapper;
     private final RedisTemplate<String, Object> redisTemplate;
+    private final RedisUtil redisUtil;
 
     @Override
     public Object findRandomShiCi() {
+        if (redisUtil.isNotEnable()) {
+            return shiCiMapper.findRandom();
+        }
         Long len = redisTemplate.opsForList().size(ApiConstant.SHI_CI);
         log.debug("redis中数据长度为" + len);
         if (len == null || len == 0) {
@@ -36,5 +41,17 @@ public class ShiCiServiceImpl extends ServiceImpl<ShiCiMapper, ShiCi> implements
             return shiCiMapper.findRandom();
         }
         return redisTemplate.opsForList().index(ApiConstant.SHI_CI, new Random().nextInt(len.intValue()));
+    }
+
+    @Override
+    public Object getShiCi(Integer id) {
+        if (redisUtil.isNotEnable()) {
+            return this.getById(id);
+        }
+        Object shiCi = redisTemplate.opsForList().index(ApiConstant.SHI_CI, id);
+        if (shiCi == null) {
+            return this.getById(id);
+        }
+        return shiCi;
     }
 }
