+++ src/main/java/com/bingchunmoli/api/utils/RedisUtil.java
@@ -1,866 +1,262 @@
 package com.bingchunmoli.api.utils;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.redis.connection.DataType;
-import org.springframework.data.redis.core.Cursor;
-import org.springframework.data.redis.core.ScanOptions;
-import org.springframework.data.redis.core.StringRedisTemplate;
-import org.springframework.data.redis.core.ZSetOperations.TypedTuple;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.data.redis.core.BoundSetOperations;
+import org.springframework.data.redis.core.HashOperations;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.core.ValueOperations;
 import org.springframework.stereotype.Component;
 
 import java.util.*;
-import java.util.Map.Entry;
 import java.util.concurrent.TimeUnit;
 
-
+/**
+ * 
+ * @author BingChunMoLi
+ */
 @Component
 @RequiredArgsConstructor
 public class RedisUtil {
-    private final StringRedisTemplate redisTemplate;
+    private final RedisTemplate redisTemplate;
+    /**
+     * Login username of the redis server.
+     */
+    @Value("${spring.redis.enable}")
+    private static boolean isEnable;
 
     /**
-     * 删除key
+     * 是否配置了redis的用户名去判断是否启用redis
+     * @return 开启状态，true为username不为null表示开启
      */
-    public void delete(String key) {
-        redisTemplate.delete(key);
+    public boolean isEnable(){
+        return isEnable;
     }
 
     /**
-     * 批量删除key
+     * 是否配置了redis的用户名去判断是否启用redis
+     * @return 开启状态，true为username为null表示不开启
      */
-    public void delete(Collection<String> keys) {
-        redisTemplate.delete(keys);
+    public boolean isNotEnable(){
+        return !isEnable;
     }
 
     /**
-     * 序列化key
+     * 缓存基本的对象，Integer、String、实体类等
+     *
+     * @param key 缓存的键值
+     * @param value 缓存的值
      */
-    public byte[] dump(String key) {
-        return redisTemplate.dump(key);
-    }
-
-    /**
-     * 是否存在key
-     */
-    public Boolean hasKey(String key) {
-        return redisTemplate.hasKey(key);
-    }
-
-    /**
-     * 设置过期时间
-     */
-    public Boolean expire(String key, long timeout, TimeUnit unit) {
-        return redisTemplate.expire(key, timeout, unit);
-    }
-
-    /**
-     * 设置过期时间
-     */
-    public Boolean expireAt(String key, Date date) {
-        return redisTemplate.expireAt(key, date);
-    }
-
-    /**
-     * 查找匹配的key
-     */
-    public Set<String> keys(String pattern) {
-        return redisTemplate.keys(pattern);
-    }
-
-    /**
-     * 将当前数据库的 key 移动到给定的数据库 db 当中
-     */
-    public Boolean move(String key, int dbIndex) {
-        return redisTemplate.move(key, dbIndex);
-    }
-
-    /**
-     * 移除 key 的过期时间，key 将持久保持
-     */
-    public Boolean persist(String key) {
-        return redisTemplate.persist(key);
-    }
-
-    /**
-     * 返回 key 的剩余的过期时间
-     */
-    public Long getExpire(String key, TimeUnit unit) {
-        return redisTemplate.getExpire(key, unit);
-    }
-
-    /**
-     * 返回 key 的剩余的过期时间
-     */
-    public Long getExpire(String key) {
-        return redisTemplate.getExpire(key);
-    }
-
-    /**
-     * 从当前数据库中随机返回一个 key
-     */
-    public String randomKey() {
-        return redisTemplate.randomKey();
-    }
-
-    /**
-     * 修改 key 的名称
-     */
-    public void rename(String oldKey, String newKey) {
-        redisTemplate.rename(oldKey, newKey);
-    }
-
-    /**
-     * 仅当 newkey 不存在时，将 oldKey 改名为 newkey
-     */
-    public Boolean renameIfAbsent(String oldKey, String newKey) {
-        return redisTemplate.renameIfAbsent(oldKey, newKey);
-    }
-
-    /**
-     * 返回 key 所储存的值的类型
-     */
-    public DataType type(String key) {
-        return redisTemplate.type(key);
-    }
-
-    /**
-     * 设置指定 key 的值
-     */
-    public void set(String key, String value) {
+    public <T> void setObject(final String key, final T value)
+    {
         redisTemplate.opsForValue().set(key, value);
     }
 
     /**
-     * 获取指定 key 的值
-     */
-    public String get(String key) {
-        return redisTemplate.opsForValue().get(key);
-    }
-
-    /**
-     * 返回 key 中字符串值的子字符
-     */
-    public String getRange(String key, long start, long end) {
-        return redisTemplate.opsForValue().get(key, start, end);
-    }
-
-    /**
-     * 将给定 key 的值设为 value ，并返回 key 的旧值(old value)
-     */
-    public String getAndSet(String key, String value) {
-        return redisTemplate.opsForValue().getAndSet(key, value);
-    }
-
-    /**
-     * 对 key 所储存的字符串值，获取指定偏移量上的位(bit)
-     */
-    public Boolean getBit(String key, long offset) {
-        return redisTemplate.opsForValue().getBit(key, offset);
-    }
-
-    /**
-     * 批量获取
-     */
-    public List<String> multiGet(Collection<String> keys) {
-        return redisTemplate.opsForValue().multiGet(keys);
-    }
-
-    /**
-     * 设置ASCII码, 字符串'a'的ASCII码是97, 转为二进制是'01100001', 此方法是将二进制第offset位值变为value
-     * <p>
-     * 值,true为1, false为0
-     */
-    public boolean setBit(String key, long offset, boolean value) {
-        return redisTemplate.opsForValue().setBit(key, offset, value);
-    }
-
-    /**
-     * 将值 value 关联到 key ，并将 key 的过期时间设为 timeout 过期时间
-     * 时间单位, 天:TimeUnit.DAYS 小时:TimeUnit.HOURS 分钟:TimeUnit.MINUTES
-     * 秒:TimeUnit.SECONDS 毫秒:TimeUnit.MILLISECONDS
-     */
-    public void setEx(String key, String value, long timeout, TimeUnit unit) {
-        redisTemplate.opsForValue().set(key, value, timeout, unit);
-    }
-
-    /**
-     * 只有在 key 不存在时设置 key 的值
-     */
-    public boolean setIfAbsent(String key, String value) {
-        return redisTemplate.opsForValue().setIfAbsent(key, value);
-    }
-
-    /**
-     * 用 value 参数覆写给定 key 所储存的字符串值，从偏移量 offset 开始
-     * 从指定位置开始覆写
-     */
-    public void setRange(String key, String value, long offset) {
-        redisTemplate.opsForValue().set(key, value, offset);
-    }
-
-    /**
-     * 获取字符串的长度
-     */
-    public Long size(String key) {
-        return redisTemplate.opsForValue().size(key);
-    }
-
-    /**
-     * 批量添加
-     */
-    public void multiSet(Map<String, String> maps) {
-        redisTemplate.opsForValue().multiSet(maps);
-    }
-
-    /**
-     * 同时设置一个或多个 key-value 对，当且仅当所有给定 key 都不存在
-     **/
-    public boolean multiSetIfAbsent(Map<String, String> maps) {
-        return redisTemplate.opsForValue().multiSetIfAbsent(maps);
-    }
-
-    /**
-     * 增加(自增长), 负数则为自减
-     */
-    public Long incrBy(String key, long increment) {
-        return redisTemplate.opsForValue().increment(key, increment);
-    }
-
-    /**
+     * 缓存基本的对象，Integer、String、实体类等
      *
+     * @param key 缓存的键值
+     * @param value 缓存的值
+     * @param timeout 时间
+     * @param timeUnit 时间颗粒度
      */
-    public Double incrByFloat(String key, double increment) {
-        return redisTemplate.opsForValue().increment(key, increment);
+    public <T> void setObject(final String key, final T value, final Integer timeout, final TimeUnit timeUnit)
+    {
+        redisTemplate.opsForValue().set(key, value, timeout, timeUnit);
     }
 
     /**
-     * 追加到末尾
-     */
-    public Integer append(String key, String value) {
-        return redisTemplate.opsForValue().append(key, value);
-    }
-
-
-    /**
-     * 获取存储在哈希表中指定字段的值
-     */
-    public Object hGet(String key, String field) {
-        return redisTemplate.opsForHash().get(key, field);
-    }
-
-    /**
-     * 获取所有给定字段的值
-     */
-    public Map<Object, Object> hGetAll(String key) {
-        return redisTemplate.opsForHash().entries(key);
-    }
-
-    /**
-     * 获取所有给定字段的值
-     */
-    public List<Object> hMultiGet(String key, Collection<Object> fields) {
-        return redisTemplate.opsForHash().multiGet(key, fields);
-    }
-
-    public void hPut(String key, String hashKey, String value) {
-        redisTemplate.opsForHash().put(key, hashKey, value);
-    }
-
-    public void hPutAll(String key, Map<String, String> maps) {
-        redisTemplate.opsForHash().putAll(key, maps);
-    }
-
-    /**
-     * 仅当hashKey不存在时才设置
-     */
-    public Boolean hPutIfAbsent(String key, String hashKey, String value) {
-        return redisTemplate.opsForHash().putIfAbsent(key, hashKey, value);
-    }
-
-    /**
-     * 删除一个或多个哈希表字段
-     */
-    public Long hDelete(String key, Object... fields) {
-        return redisTemplate.opsForHash().delete(key, fields);
-    }
-
-    /**
-     * 查看哈希表 key 中，指定的字段是否存在
-     */
-    public boolean hExists(String key, String field) {
-        return redisTemplate.opsForHash().hasKey(key, field);
-    }
-
-    /**
-     * 为哈希表 key 中的指定字段的整数值加上增量 increment
-     */
-    public Long hIncrBy(String key, Object field, long increment) {
-        return redisTemplate.opsForHash().increment(key, field, increment);
-    }
-
-    /**
-     * 为哈希表 key 中的指定字段的整数值加上增量 increment
-     */
-    public Double hIncrByFloat(String key, Object field, double delta) {
-        return redisTemplate.opsForHash().increment(key, field, delta);
-    }
-
-    /**
-     * 获取所有哈希表中的字段
-     */
-    public Set<Object> hKeys(String key) {
-        return redisTemplate.opsForHash().keys(key);
-    }
-
-    /**
-     * 获取哈希表中字段的数量
-     */
-    public Long hSize(String key) {
-        return redisTemplate.opsForHash().size(key);
-    }
-
-    /**
-     * 获取哈希表中所有值
-     */
-    public List<Object> hValues(String key) {
-        return redisTemplate.opsForHash().values(key);
-    }
-
-    /**
-     * 迭代哈希表中的键值对
-     */
-    public Cursor<Entry<Object, Object>> hScan(String key, ScanOptions options) {
-        return redisTemplate.opsForHash().scan(key, options);
-    }
-
-
-    /**
-     * 通过索引获取列表中的元素
-     */
-    public String lIndex(String key, long index) {
-        return redisTemplate.opsForList().index(key, index);
-    }
-
-    /**
-     * 获取列表指定范围内的元素
-     * <p>
-     * 开始位置, 0是开始位置
-     * 结束位置, -1返回所有
-     */
-    public List<String> lRange(String key, long start, long end) {
-        return redisTemplate.opsForList().range(key, start, end);
-    }
-
-    /**
-     * 存储在list头部
-     */
-    public Long lLeftPush(String key, String value) {
-        return redisTemplate.opsForList().leftPush(key, value);
-    }
-
-    /**
+     * 设置有效时间
      *
+     * @param key Redis键
+     * @param timeout 超时时间
+     * @return true=设置成功；false=设置失败
      */
-    public Long lLeftPushAll(String key, String... value) {
-        return redisTemplate.opsForList().leftPushAll(key, value);
+    public boolean expire(final String key, final long timeout)
+    {
+        return expire(key, timeout, TimeUnit.SECONDS);
     }
 
     /**
+     * 设置有效时间
      *
+     * @param key Redis键
+     * @param timeout 超时时间
+     * @param unit 时间单位
+     * @return true=设置成功；false=设置失败
      */
-    public Long lLeftPushAll(String key, Collection<String> value) {
-        return redisTemplate.opsForList().leftPushAll(key, value);
+    public boolean expire(final String key, final long timeout, final TimeUnit unit)
+    {
+        return redisTemplate.expire(key, timeout, unit);
     }
 
     /**
-     * 当list存在的时候才加入
+     * 获得缓存的基本对象。
+     *
+     * @param key 缓存键值
+     * @return 缓存键值对应的数据
      */
-    public Long lLeftPushIfPresent(String key, String value) {
-        return redisTemplate.opsForList().leftPushIfPresent(key, value);
+    public <T> T getObject(final String key)
+    {
+        ValueOperations<String, T> operation = redisTemplate.opsForValue();
+        return operation.get(key);
     }
 
     /**
-     * 如果pivot存在,再pivot前面添加
+     * 删除单个对象
+     *
+     * @param key
      */
-    public Long lLeftPush(String key, String pivot, String value) {
-        return redisTemplate.opsForList().leftPush(key, pivot, value);
-    }
-
-
-    public Long lRightPush(String key, String value) {
-        return redisTemplate.opsForList().rightPush(key, value);
-    }
-
-
-    public Long lRightPushAll(String key, String... value) {
-        return redisTemplate.opsForList().rightPushAll(key, value);
-    }
-
-
-    public Long lRightPushAll(String key, Collection<String> value) {
-        return redisTemplate.opsForList().rightPushAll(key, value);
+    public boolean deleteObject(final String key)
+    {
+        return redisTemplate.delete(key);
     }
 
     /**
-     * 为已存在的列表添加值
+     * 删除集合对象
+     *
+     * @param collection 多个对象
+     * @return
      */
-    public Long lRightPushIfPresent(String key, String value) {
-        return redisTemplate.opsForList().rightPushIfPresent(key, value);
+    public long deleteObject(final Collection collection)
+    {
+        return redisTemplate.delete(collection);
     }
 
     /**
-     * 在pivot元素的右边添加值
+     * 缓存List数据
+     *
+     * @param key 缓存的键值
+     * @param dataList 待缓存的List数据
+     * @return 缓存的对象
      */
-    public Long lRightPush(String key, String pivot, String value) {
-        return redisTemplate.opsForList().rightPush(key, pivot, value);
+    public <T> long setList(final String key, final List<T> dataList)
+    {
+        Long count = redisTemplate.opsForList().rightPushAll(key, dataList);
+        return count == null ? 0 : count;
     }
 
     /**
-     * 通过索引设置列表元素的值
-     * <p>
-     * 位置
+     * 获得缓存的list对象
+     *
+     * @param key 缓存的键值
+     * @return 缓存键值对应的数据
      */
-    public void lSet(String key, long index, String value) {
-        redisTemplate.opsForList().set(key, index, value);
+    public <T> List<T> getList(final String key)
+    {
+        return redisTemplate.opsForList().range(key, 0, -1);
     }
 
     /**
-     * 移出并获取列表的第一个元素
+     * 缓存Set
+     *
+     * @param key 缓存键值
+     * @param dataSet 缓存的数据
+     * @return 缓存数据的对象
      */
-    public String lLeftPop(String key) {
-        return redisTemplate.opsForList().leftPop(key);
+    public <T> BoundSetOperations<String, T> setSet(final String key, final Set<T> dataSet)
+    {
+        BoundSetOperations<String, T> setOperation = redisTemplate.boundSetOps(key);
+        Iterator<T> it = dataSet.iterator();
+        while (it.hasNext())
+        {
+            setOperation.add(it.next());
+        }
+        return setOperation;
     }
 
     /**
-     * 移出并获取列表的第一个元素， 如果列表没有元素会阻塞列表直到等待超时或发现可弹出元素为止
-     * <p>
-     * 等待时间
-     * 时间单位
+     * 获得缓存的set
+     *
+     * @param key
+     * @return
      */
-    public String lBLeftPop(String key, long timeout, TimeUnit unit) {
-        return redisTemplate.opsForList().leftPop(key, timeout, unit);
-    }
-
-    /**
-     * 移除并获取列表最后一个元素
-     */
-    public String lRightPop(String key) {
-        return redisTemplate.opsForList().rightPop(key);
-    }
-
-    /**
-     * 移出并获取列表的最后一个元素， 如果列表没有元素会阻塞列表直到等待超时或发现可弹出元素为止
-     * <p>
-     * 等待时间
-     * 时间单位
-     */
-    public String lBRightPop(String key, long timeout, TimeUnit unit) {
-        return redisTemplate.opsForList().rightPop(key, timeout, unit);
-    }
-
-    /**
-     * 移除列表的最后一个元素，并将该元素添加到另一个列表并返回
-     */
-    public String lRightPopAndLeftPush(String sourceKey, String destinationKey) {
-        return redisTemplate.opsForList().rightPopAndLeftPush(sourceKey,
-                destinationKey);
-    }
-
-    /**
-     * 从列表中弹出一个值，将弹出的元素插入到另外一个列表中并返回它； 如果列表没有元素会阻塞列表直到等待超时或发现可弹出元素为止
-     */
-    public String lBRightPopAndLeftPush(String sourceKey, String destinationKey,
-                                        long timeout, TimeUnit unit) {
-        return redisTemplate.opsForList().rightPopAndLeftPush(sourceKey,
-                destinationKey, timeout, unit);
-    }
-
-    /**
-     * 删除集合中值等于value得元素
-     * <p>
-     * index=0, 删除所有值等于value的元素; index>0, 从头部开始删除第一个值等于value的元素;
-     * index<0, 从尾部开始删除第一个值等于value的元素;
-     */
-    public Long lRemove(String key, long index, String value) {
-        return redisTemplate.opsForList().remove(key, index, value);
-    }
-
-    /**
-     * 裁剪list
-     */
-    public void lTrim(String key, long start, long end) {
-        redisTemplate.opsForList().trim(key, start, end);
-    }
-
-    /**
-     * 获取列表长度
-     */
-    public Long lLen(String key) {
-        return redisTemplate.opsForList().size(key);
-    }
-
-
-    /**
-     * set添加元素
-     */
-    public Long sAdd(String key, String... values) {
-        return redisTemplate.opsForSet().add(key, values);
-    }
-
-    /**
-     * set移除元素
-     */
-    public Long sRemove(String key, Object... values) {
-        return redisTemplate.opsForSet().remove(key, values);
-    }
-
-    /**
-     * 移除并返回集合的一个随机元素
-     */
-    public String sPop(String key) {
-        return redisTemplate.opsForSet().pop(key);
-    }
-
-    /**
-     * 将元素value从一个集合移到另一个集合
-     */
-    public Boolean sMove(String key, String value, String destKey) {
-        return redisTemplate.opsForSet().move(key, value, destKey);
-    }
-
-    /**
-     * 获取集合的大小
-     */
-    public Long sSize(String key) {
-        return redisTemplate.opsForSet().size(key);
-    }
-
-    /**
-     * 判断集合是否包含value
-     */
-    public Boolean sIsMember(String key, Object value) {
-        return redisTemplate.opsForSet().isMember(key, value);
-    }
-
-    /**
-     * 获取两个集合的交集
-     */
-    public Set<String> sIntersect(String key, String otherKey) {
-        return redisTemplate.opsForSet().intersect(key, otherKey);
-    }
-
-    /**
-     * 获取key集合与多个集合的交集
-     */
-    public Set<String> sIntersect(String key, Collection<String> otherKeys) {
-        return redisTemplate.opsForSet().intersect(key, otherKeys);
-    }
-
-    /**
-     * key集合与otherKey集合的交集存储到destKey集合中
-     */
-    public Long sIntersectAndStore(String key, String otherKey, String destKey) {
-        return redisTemplate.opsForSet().intersectAndStore(key, otherKey,
-                destKey);
-    }
-
-    /**
-     * key集合与多个集合的交集存储到destKey集合中
-     */
-    public Long sIntersectAndStore(String key, Collection<String> otherKeys,
-                                   String destKey) {
-        return redisTemplate.opsForSet().intersectAndStore(key, otherKeys,
-                destKey);
-    }
-
-    /**
-     * 获取两个集合的并集
-     */
-    public Set<String> sUnion(String key, String otherKeys) {
-        return redisTemplate.opsForSet().union(key, otherKeys);
-    }
-
-    /**
-     * 获取key集合与多个集合的并集
-     */
-    public Set<String> sUnion(String key, Collection<String> otherKeys) {
-        return redisTemplate.opsForSet().union(key, otherKeys);
-    }
-
-    /**
-     * key集合与otherKey集合的并集存储到destKey中
-     */
-    public Long sUnionAndStore(String key, String otherKey, String destKey) {
-        return redisTemplate.opsForSet().unionAndStore(key, otherKey, destKey);
-    }
-
-    /**
-     * key集合与多个集合的并集存储到destKey中
-     */
-    public Long sUnionAndStore(String key, Collection<String> otherKeys,
-                               String destKey) {
-        return redisTemplate.opsForSet().unionAndStore(key, otherKeys, destKey);
-    }
-
-    /**
-     * 获取两个集合的差集
-     */
-    public Set<String> sDifference(String key, String otherKey) {
-        return redisTemplate.opsForSet().difference(key, otherKey);
-    }
-
-    /**
-     * 获取key集合与多个集合的差集
-     */
-    public Set<String> sDifference(String key, Collection<String> otherKeys) {
-        return redisTemplate.opsForSet().difference(key, otherKeys);
-    }
-
-    /**
-     * key集合与otherKey集合的差集存储到destKey中
-     */
-    public Long sDifference(String key, String otherKey, String destKey) {
-        return redisTemplate.opsForSet().differenceAndStore(key, otherKey,
-                destKey);
-    }
-
-    /**
-     * key集合与多个集合的差集存储到destKey中
-     */
-    public Long sDifference(String key, Collection<String> otherKeys,
-                            String destKey) {
-        return redisTemplate.opsForSet().differenceAndStore(key, otherKeys,
-                destKey);
-    }
-
-    /**
-     * 获取集合所有元素
-     */
-    public Set<String> setMembers(String key) {
+    public <T> Set<T> getSet(final String key)
+    {
         return redisTemplate.opsForSet().members(key);
     }
 
     /**
-     * 随机获取集合中的一个元素
+     * 缓存Map
+     *
+     * @param key
+     * @param dataMap
      */
-    public String sRandomMember(String key) {
-        return redisTemplate.opsForSet().randomMember(key);
+    public <T> void setMap(final String key, final Map<String, T> dataMap)
+    {
+        if (dataMap != null) {
+            redisTemplate.opsForHash().putAll(key, dataMap);
+        }
     }
 
     /**
-     * 随机获取集合中count个元素
+     * 获得缓存的Map
+     *
+     * @param key
+     * @return
      */
-    public List<String> sRandomMembers(String key, long count) {
-        return redisTemplate.opsForSet().randomMembers(key, count);
+    public <T> Map<String, T> getMap(final String key)
+    {
+        return redisTemplate.opsForHash().entries(key);
     }
 
     /**
-     * 随机获取集合中count个元素并且去除重复的
+     * 往Hash中存入数据
+     *
+     * @param key Redis键
+     * @param hKey Hash键
+     * @param value 值
      */
-    public Set<String> sDistinctRandomMembers(String key, long count) {
-        return redisTemplate.opsForSet().distinctRandomMembers(key, count);
-    }
-
-
-    public Cursor<String> sScan(String key, ScanOptions options) {
-        return redisTemplate.opsForSet().scan(key, options);
-    }
-
-
-    /**
-     * 添加元素,有序集合是按照元素的score值由小到大排列
-     */
-    public Boolean zAdd(String key, String value, double score) {
-        return redisTemplate.opsForZSet().add(key, value, score);
-    }
-
-
-    public Long zAdd(String key, Set<TypedTuple<String>> values) {
-        return redisTemplate.opsForZSet().add(key, values);
-    }
-
-
-    public Long zRemove(String key, Object... values) {
-        return redisTemplate.opsForZSet().remove(key, values);
+    public <T> void setMapValue(final String key, final String hKey, final T value)
+    {
+        redisTemplate.opsForHash().put(key, hKey, value);
     }
 
     /**
-     * 增加元素的score值，并返回增加后的值
+     * 获取Hash中的数据
+     *
+     * @param key Redis键
+     * @param hKey Hash键
+     * @return Hash中的对象
      */
-    public Double zIncrementScore(String key, String value, double delta) {
-        return redisTemplate.opsForZSet().incrementScore(key, value, delta);
+    public <T> T getMapValue(final String key, final String hKey)
+    {
+        HashOperations<String, String, T> opsForHash = redisTemplate.opsForHash();
+        return opsForHash.get(key, hKey);
     }
 
     /**
-     * 返回元素在集合的排名,有序集合是按照元素的score值由小到大排列
+     * 删除Hash中的数据
+     *
+     * @param key
+     * @param hKey
      */
-    public Long zRank(String key, Object value) {
-        return redisTemplate.opsForZSet().rank(key, value);
+    public void delMapValue(final String key, final String hKey)
+    {
+        HashOperations hashOperations = redisTemplate.opsForHash();
+        hashOperations.delete(key, hKey);
     }
 
     /**
-     * 返回元素在集合的排名,按元素的score值由大到小排列
+     * 获取多个Hash中的数据
+     *
+     * @param key Redis键
+     * @param hKeys Hash键集合
+     * @return Hash对象集合
      */
-    public Long zReverseRank(String key, Object value) {
-        return redisTemplate.opsForZSet().reverseRank(key, value);
+    public <T> List<T> getMultiMapValue(final String key, final Collection<Object> hKeys)
+    {
+        return redisTemplate.opsForHash().multiGet(key, hKeys);
     }
 
     /**
-     * 获取集合的元素, 从小到大排序
-     * 开始位置
-     * 结束位置, -1查询所有
+     * 获得缓存的基本对象列表
+     *
+     * @param pattern 字符串前缀
+     * @return 对象列表
      */
-    public Set<String> zRange(String key, long start, long end) {
-        return redisTemplate.opsForZSet().range(key, start, end);
-    }
-
-    /**
-     * 获取集合元素, 并且把score值也获取
-     */
-    public Set<TypedTuple<String>> zRangeWithScores(String key, long start,
-                                                    long end) {
-        return redisTemplate.opsForZSet().rangeWithScores(key, start, end);
-    }
-
-    /**
-     * 根据Score值查询集合元素
-     * 最小值
-     * 最大值
-     */
-    public Set<String> zRangeByScore(String key, double min, double max) {
-        return redisTemplate.opsForZSet().rangeByScore(key, min, max);
-    }
-
-    /**
-     * 根据Score值查询集合元素, 从小到大排序
-     * 最小值
-     * 最大值
-     */
-    public Set<TypedTuple<String>> zRangeByScoreWithScores(String key,
-                                                           double min, double max) {
-        return redisTemplate.opsForZSet().rangeByScoreWithScores(key, min, max);
-    }
-
-
-    public Set<TypedTuple<String>> zRangeByScoreWithScores(String key,
-                                                           double min, double max, long start, long end) {
-        return redisTemplate.opsForZSet().rangeByScoreWithScores(key, min, max,
-                start, end);
-    }
-
-    /**
-     * 获取集合的元素, 从大到小排序
-     */
-    public Set<String> zReverseRange(String key, long start, long end) {
-        return redisTemplate.opsForZSet().reverseRange(key, start, end);
-    }
-
-    /**
-     * 获取集合的元素, 从大到小排序, 并返回score值
-     */
-    public Set<TypedTuple<String>> zReverseRangeWithScores(String key,
-                                                           long start, long end) {
-        return redisTemplate.opsForZSet().reverseRangeWithScores(key, start,
-                end);
-    }
-
-    /**
-     * 根据Score值查询集合元素, 从大到小排序
-     */
-    public Set<String> zReverseRangeByScore(String key, double min,
-                                            double max) {
-        return redisTemplate.opsForZSet().reverseRangeByScore(key, min, max);
-    }
-
-    /**
-     * 根据Score值查询集合元素, 从大到小排序
-     */
-    public Set<TypedTuple<String>> zReverseRangeByScoreWithScores(
-            String key, double min, double max) {
-        return redisTemplate.opsForZSet().reverseRangeByScoreWithScores(key,
-                min, max);
-    }
-
-    public Set<String> zReverseRangeByScore(String key, double min,
-                                            double max, long start, long end) {
-        return redisTemplate.opsForZSet().reverseRangeByScore(key, min, max,
-                start, end);
-    }
-
-    /**
-     * 根据score值获取集合元素数量
-     */
-    public Long zCount(String key, double min, double max) {
-        return redisTemplate.opsForZSet().count(key, min, max);
-    }
-
-    /**
-     * 获取集合大小
-     */
-    public Long zSize(String key) {
-        return redisTemplate.opsForZSet().size(key);
-    }
-
-    /**
-     * 获取集合大小
-     */
-    public Long zZCard(String key) {
-        return redisTemplate.opsForZSet().zCard(key);
-    }
-
-    /**
-     * 获取集合中value元素的score值
-     */
-    public Double zScore(String key, Object value) {
-        return redisTemplate.opsForZSet().score(key, value);
-    }
-
-    /**
-     * 移除指定索引位置的成员
-     */
-    public Long zRemoveRange(String key, long start, long end) {
-        return redisTemplate.opsForZSet().removeRange(key, start, end);
-    }
-
-    /**
-     * 根据指定的score值的范围来移除成员
-     */
-    public Long zRemoveRangeByScore(String key, double min, double max) {
-        return redisTemplate.opsForZSet().removeRangeByScore(key, min, max);
-    }
-
-    /**
-     * 获取key和otherKey的并集并存储在destKey中
-     */
-    public Long zUnionAndStore(String key, String otherKey, String destKey) {
-        return redisTemplate.opsForZSet().unionAndStore(key, otherKey, destKey);
-    }
-
-    public Long zUnionAndStore(String key, Collection<String> otherKeys,
-                               String destKey) {
-        return redisTemplate.opsForZSet()
-                .unionAndStore(key, otherKeys, destKey);
-    }
-
-    /**
-     * 交集
-     */
-    public Long zIntersectAndStore(String key, String otherKey,
-                                   String destKey) {
-        return redisTemplate.opsForZSet().intersectAndStore(key, otherKey,
-                destKey);
-    }
-
-    /**
-     * 交集
-     */
-    public Long zIntersectAndStore(String key, Collection<String> otherKeys,
-                                   String destKey) {
-        return redisTemplate.opsForZSet().intersectAndStore(key, otherKeys,
-                destKey);
-    }
-
-    public Cursor<TypedTuple<String>> zScan(String key, ScanOptions options) {
-        return redisTemplate.opsForZSet().scan(key, options);
+    public Collection<String> keys(final String pattern)
+    {
+        return redisTemplate.keys(pattern);
     }
 }
