+++ src/main/java/com/bingchunmoli/api/weather/service/impl/WeatherServiceImpl.java
@@ -5,7 +5,7 @@ import cn.hutool.http.HttpUtil;
 import com.alibaba.fastjson.JSON;
 import com.bingchunmoli.api.properties.ApiKeyProperties;
 import com.bingchunmoli.api.utils.IntegerUtil;
-import com.bingchunmoli.api.utils.RedisUtil;
+import com.bingchunmoli.api.utils.StringRedisUtil;
 import com.bingchunmoli.api.weather.bean.WeatherVO;
 import com.bingchunmoli.api.weather.bean.enums.WeatherCacheKey;
 import com.bingchunmoli.api.weather.service.WeatherService;
@@ -25,10 +25,13 @@ import java.util.concurrent.TimeUnit;
 @RequiredArgsConstructor
 public class WeatherServiceImpl implements WeatherService {
     private final ApiKeyProperties apiKeyProperties;
-    private final RedisUtil redisUtil;
+    private final StringRedisUtil stringRedisUtil;
 
     @Override
     public String getWeatherByDay(Integer day, String location) throws UnsupportedEncodingException {
+        if (stringRedisUtil.isNotEnable()) {
+            return "未开通此功能";
+        }
         if (location.contains(StrPool.COMMA) || IntegerUtil.isInteget(location)) {
             // 按经维度查询 或者 id查询
             return getWeatherByDayCommon(day, location);
@@ -39,8 +42,11 @@ public class WeatherServiceImpl implements WeatherService {
 
     @Override
     public String getWeatherByNow(String address) throws UnsupportedEncodingException {
+        if (stringRedisUtil.isNotEnable()) {
+            return "未开通此功能";
+        }
         String redisCacheKey = new StringJoiner(":", WeatherCacheKey.BY_NOW.getKey(), ":" + address).toString();
-        String redisCache = redisUtil.get(redisCacheKey);
+        String redisCache = stringRedisUtil.get(redisCacheKey);
         return Optional.ofNullable(redisCache).orElse(doGetWeatherByNow(redisCacheKey, address));
     }
 
@@ -55,7 +61,7 @@ public class WeatherServiceImpl implements WeatherService {
     private String getWeatherByDayCommon(Integer day, String location) throws UnsupportedEncodingException {
         String redisCacheKey = new StringJoiner(":", WeatherCacheKey.BY_DAY.getKey(), ":" + location)
                 .add(String.valueOf(day)).toString();
-        String redisCache = redisUtil.get(redisCacheKey);
+        String redisCache = stringRedisUtil.get(redisCacheKey);
         return Optional.ofNullable(redisCache).orElse(doGetWeatherByDay(redisCacheKey, day, location));
     }
 
@@ -78,7 +84,7 @@ public class WeatherServiceImpl implements WeatherService {
                 "&location=" +
                 getLocationId(location);
         String res = HttpUtil.get(joiner);
-        redisUtil.setEx(redisCacheKey, res, 12, TimeUnit.HOURS);
+        stringRedisUtil.setEx(redisCacheKey, res, 12, TimeUnit.HOURS);
         return res;
     }
 
@@ -99,7 +105,7 @@ public class WeatherServiceImpl implements WeatherService {
                 "&location=" +
                 getLocationId(location);
         String res = HttpUtil.get(requestUrl);
-        redisUtil.setEx(redisCacheKey, res, 1, TimeUnit.HOURS);
+        stringRedisUtil.setEx(redisCacheKey, res, 1, TimeUnit.HOURS);
         return res;
     }
 
@@ -112,7 +118,7 @@ public class WeatherServiceImpl implements WeatherService {
      */
     private String getLocationId(String location) throws UnsupportedEncodingException {
         String redisCacheKey = new StringJoiner(":", WeatherCacheKey.LOOKUP.getKey(), location).toString();
-        String redisCache = redisUtil.get(redisCacheKey);
+        String redisCache = stringRedisUtil.get(redisCacheKey);
         String res = Optional.ofNullable(redisCache).orElse(doGetLocationId(redisCacheKey, location));
         return JSON.parseObject(res, WeatherVO.class).getLocation().get(0).getId();
     }
@@ -133,7 +139,7 @@ public class WeatherServiceImpl implements WeatherService {
                 "&location=" +
                 URLEncoder.encode(location, "utf-8");
         String res = HttpUtil.get(requestUrl);
-        redisUtil.setEx(redisCacheKey, res, 24, TimeUnit.HOURS);
+        stringRedisUtil.setEx(redisCacheKey, res, 24, TimeUnit.HOURS);
         return res;
     }
