+++ src/main/java/com/bingchunmoli/api/yiyan/mapper/YiYanMapper.java
@@ -20,6 +20,6 @@ public interface YiYanMapper extends BaseMapper<YiYan> {
      *
      * @return com.bingchunmoli.api.entity.Yiyan
      */
-    @Select("SELECT * FROM yi_yan as t1 WHERE t1.id>=(RAND()*(SELECT MAX(id) FROM yiyan))LIMIT 1")
+    @Select("SELECT * FROM yi_yan as t1 WHERE t1.id>=(RAND()*(SELECT MAX(id) FROM yi_yan))LIMIT 1")
     YiYan findRandom();
 }
