+++ src/main/java/com/bingchunmoli/api/yiyan/service/IYiYanService.java
@@ -13,4 +13,6 @@ public interface IYiYanService extends IService<YiYan> {
      * @return 一言对象
      */
     Object findRandomYiYan();
+
+    Object getYiYan(Integer id);
 }
