+++ src/main/resources/application-h2.yml
@@ -1,0 +1,58 @@
+#INIT=runscript from 'classpath:db/schema-h2.sql'
+spring:
+  datasource:
+    driver-class-name: org.h2.Driver
+    url: jdbc:h2:file:~/.api/api.db;MODE=MySQL;AUTO_SERVER=TRUE;DATABASE_TO_LOWER=TRUE;database_to_upper=false;
+    username: root
+    password: root
+  sql:
+    init:
+      schema-locations: classpath:db/schema-h2.sql
+      data-locations: classpath:/db/data-h2.sql
+      username: root
+      password: root
+      encoding: UTF-8
+      platform: h2
+      mode: ALWAYS
+      continue-on-error: true
+  mail:
+    enable: false
+    host:
+    port: 465
+    username:
+    password:
+    default-encoding: UTF-8
+    protocol: smtps
+    defaultTo:
+  redis:
+    enable: false
+    database: 1
+    password:
+    port: 6379
+    host: localhost
+management:
+  #  endpoints:
+  #    web:
+  #      exposure:
+  #        include: '*'
+  endpoint:
+    health:
+      show-details: WHEN_AUTHORIZED
+springfox:
+  documentation:
+    swagger-ui:
+      enabled: false
+
+moli:
+  apiConfig:
+    weatherKey:
+    weatherUri:
+    weatherGeoUri:
+    serverSauceKey:
+    pcPath:
+    mobilePath:
+    path1080:
+logging:
+  level:
+    com.bingchunmoli: info
+    com.baomidou.example.mapper: info
